+++ example/nestjs-vue3-ssr/.eslintrc.js
@@ -2,8 +2,5 @@ module.exports = {
   root: true,
   extends: [
     'standard-vue-ts'
-  ],
-  parserOptions: {
-    project: './tsconfig.lint.json'
-  }
+  ]
 }