+++ packages/plugin-nestjs/src/start.ts
@@ -1,9 +1,10 @@
 import { exec } from 'child_process'
-import { logGreen } from 'ssr-server-utils'
+import { logGreen, loadConfig } from 'ssr-server-utils'
 
 const spinner = require('ora')('starting ')
 
 const start = () => {
+  const config = loadConfig()
   spinner.start()
   const { stdout, stderr } = exec('npx nest start --watch', {} /* options, [optional] */)
   stdout?.on('data', function (data) {
@@ -11,7 +12,7 @@ const start = () => {
     if (data.match('Nest application successfully started')) {
       spinner.stop()
       const https = process.env.HTTPS
-      logGreen(`Server is listening on ${https ? 'https' : 'http'}://localhost:3000`)
+      logGreen(`Server is listening on ${https ? 'https' : 'http'}://localhost:${config.serverPort}`)
     }
   })
   stderr?.on('data', function (data) {