+++ packages/plugin-vue/src/entry/client-entry.ts
@@ -1,14 +1,29 @@
 import Vue from 'vue'
+import { Store } from 'vuex'
+import { Route } from 'vue-router'
 import { findRoute } from 'ssr-client-utils'
-import { FeRouteItem } from 'ssr-types'
 // @ts-expect-error
-import feRoutes from 'ssr-temporary-routes'
+import * as Routes from 'ssr-temporary-routes'
+import { ESMFetch, RoutesType, IClientFeRouteItem } from './interface'
 import { createRouter } from './router'
 import { createStore } from './store'
 
 declare const module: any
+const { FeRoutes, App, layoutFetch } = Routes as RoutesType
 
-const App = feRoutes[0].App
+async function getAsyncCombineData (fetch: ESMFetch | undefined, store: Store<any>, router: Route) {
+  let layoutFetchData = {}
+  let fetchData = {}
+  if (layoutFetch) {
+    layoutFetchData = await layoutFetch({ store, router })
+  }
+  if (fetch) {
+    const fetchFn = await fetch()
+    fetchData = await fetchFn.default({ store, router })
+  }
+  return Object.assign({}, layoutFetchData ?? {}, fetchData ?? {})
+}
+
 const clientRender = async () => {
   const store = createStore()
   const router = createRouter()
@@ -27,25 +42,23 @@ const clientRender = async () => {
   router.onReady(async () => {
     if (!window.__USE_SSR__) {
       // 如果是 csr 模式 则需要客户端获取首页需要的数据
-      const route = findRoute<FeRouteItem>(feRoutes, location.pathname)
+      const route = findRoute<IClientFeRouteItem>(FeRoutes, location.pathname)
       const { fetch } = route
-      if (fetch) {
-        await fetch({ store, router: router.currentRoute })
-      }
+      await getAsyncCombineData(fetch, store, router.currentRoute)
     }
     router.beforeResolve(async (to, from, next) => {
       // 找到要进入的组件并提前执行 fetch 函数
-      const route = findRoute<FeRouteItem>(feRoutes, to.path)
-      if (route.fetch) {
-        await route.fetch({ store, router: to })
-      }
+      const route = findRoute<IClientFeRouteItem>(FeRoutes, to.path)
+      const { fetch } = route
+      await getAsyncCombineData(fetch, store, to)
+
       next()
     })
     app.$mount('#app', !!window.__USE_SSR__) // 这里需要做判断 ssr/csr 来为 true/false
   })
 
-  if (process.env.NODE_ENV === 'development' && module.hot) {
-    module.hot.accept()
+  if (!window.__USE_VITE__) {
+    module?.hot?.accept?.() // webpack 场景下的 hmr
   }
 }