+++ packages/plugin-vue3/src/entry/client-entry.ts
@@ -1,13 +1,29 @@
-import { h, createApp } from 'vue'
+import { h, createApp, reactive } from 'vue'
+import { Store } from 'vuex'
+import { RouteLocationNormalizedLoaded } from 'vue-router'
 import { findRoute } from 'ssr-client-utils'
-import { ESMFeRouteItem } from 'ssr-types'
 import { createRouter } from './router'
 import { createStore } from './store'
+import { ESMFetch, IClientFeRouteItem, RoutesType } from './interface'
 
 // @ts-expect-error
-import feRoutes from 'ssr-temporary-routes'
+import * as Routes from 'ssr-temporary-routes'
 
+const { FeRoutes, App, layoutFetch } = Routes as RoutesType
 declare const module: any
+
+async function getAsyncCombineData (fetch: ESMFetch | undefined, store: Store<any>, router: RouteLocationNormalizedLoaded) {
+  let layoutFetchData = {}
+  let fetchData = {}
+  if (layoutFetch) {
+    layoutFetchData = await layoutFetch({ store, router })
+  }
+  if (fetch) {
+    const fetchFn = await fetch()
+    fetchData = await fetchFn.default({ store, router })
+  }
+  return Object.assign({}, layoutFetchData ?? {}, fetchData ?? {})
+}
 
 const clientRender = async () => {
   const store = createStore()
@@ -16,38 +32,38 @@ const clientRender = async () => {
   if (window.__INITIAL_DATA__) {
     store.replaceState(window.__INITIAL_DATA__)
   }
-  const App = await feRoutes[0].App()
+
+  const asyncData = reactive({
+    value: window.__INITIAL_DATA__ ?? {}
+  })
   const app = createApp({
-    render: () => h(App.default)
+    render: () => h(App, {
+      asyncData
+    })
   })
 
   app.use(store)
   app.use(router)
 
-  window.__VUE_APP__ = app
-
   await router.isReady()
+  router.beforeResolve(async (to, from, next) => {
+    // 找到要进入的组件并提前执行 fetch 函数
+    const { fetch } = findRoute<IClientFeRouteItem>(FeRoutes, to.path)
+    const combineAysncData = await getAsyncCombineData(fetch, store, to)
+    asyncData.value = Object.assign(asyncData.value, combineAysncData)
+    next()
+  })
 
   if (!window.__USE_SSR__) {
     // 如果是 csr 模式 则需要客户端获取首页需要的数据
-    const route = findRoute<ESMFeRouteItem>(feRoutes, location.pathname)
-    const { fetch } = route
-
-    if (fetch) {
-      const fetchFn = await fetch()
-      await fetchFn.default({ store, router: router.currentRoute })
-    }
+    const { fetch } = findRoute<IClientFeRouteItem>(FeRoutes, location.pathname)
+    const combineAysncData = await getAsyncCombineData(fetch, store, router.currentRoute.value)
+    asyncData.value = Object.assign(asyncData.value, combineAysncData)
   }
 
-  router.beforeResolve(async (to, from, next) => {
-    // 找到要进入的组件并提前执行 fetch 函数
-    const route = findRoute<ESMFeRouteItem>(feRoutes, to.path)
-    if (route.fetch) {
-      const fetchFn = await route.fetch()
-      await fetchFn.default({ store, router: to })
-    }
-    next()
-  })
+  window.__VUE_APP__ = app
+  window.__VUE_ROUTER__ = router
+
   app.mount('#app', !!window.__USE_SSR__) // 这里需要做判断 ssr/csr 来为 true/false
   if (!window.__USE_VITE__) {
     module?.hot?.accept?.() // webpack 场景下的 hmr