+++ packages/server-utils/src/cwd.ts
@@ -51,7 +51,14 @@ const accessFile = async (file: string) => {
 const checkVite = async () => {
   const result = await accessFile(resolve(getCwd(), './node_modules/vite/package.json'))
   if (!result) {
-    console.log('当前项目缺少 vite 依赖，请根据实际技术栈安装 vite @vitejs/plugin-vue 或 其他对应插件')
+    const version = require(resolve(getCwd(), './package.json')).dependencies.vue
+    let plugin = ''
+    if (version) {
+      plugin = /^.?3/.test(version) ? '@vitejs/plugin-vue' : 'vite-plugin-vue2'
+    } else {
+      plugin = '@vitejs/plugin-react-refresh'
+    }
+    console.log(`当前项目缺少 vite 依赖，请根据实际技术栈安装 vite ${plugin} 或 其他对应插件`)
     return false
   }
   return true
@@ -61,9 +68,15 @@ const copyViteConfig = async () => {
   // 如果当前项目没有 vite.config 则复制默认的文件
   const result = await accessFile(resolve(getCwd(), './vite.config.js'))
   if (!result) {
+    const version = require(resolve(getCwd(), './package.json')).dependencies.vue
     console.log('检测到当前目录缺少 vite.config.js 文件，自动创建默认模版')
-
-    await promises.copyFile(resolve(getCwd(), './node_modules/ssr-plugin-vue3/src/config/vite.config.tpl'), resolve(getCwd(), './vite.config.js'))
+    let folder = ''
+    if (version) {
+      folder = /^.?3/.test(version) ? 'ssr-plugin-vue3' : 'ssr-plugin-vue'
+    } else {
+      folder = 'ssr-plugin-react'
+    }
+    await promises.copyFile(resolve(getCwd(), `./node_modules/${folder}/src/config/vite.config.tpl`), resolve(getCwd(), './vite.config.js'))
   }
 }