+++ packages/core-vue/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.5.15](https://github.com/ykfe/ssr/compare/v5.5.14...v5.5.15) (2021-04-19)
+
+**Note:** Version bump only for package ssr-core-vue
+
+
+
+
+
+## [5.5.14](https://github.com/ykfe/ssr/compare/v5.5.13...v5.5.14) (2021-04-17)
+
+**Note:** Version bump only for package ssr-core-vue
+
+
+
+
+
+## [5.5.13](https://github.com/ykfe/ssr/compare/v5.5.12...v5.5.13) (2021-04-15)
+
+**Note:** Version bump only for package ssr-core-vue
+
+
+
+
+
+## [5.5.11](https://github.com/ykfe/ssr/compare/v5.5.10...v5.5.11) (2021-04-14)
+
+
+### Bug Fixes
+
+* core-vue dependencies ([79dff40](https://github.com/ykfe/ssr/commit/79dff40862281afe92213a40ca2724999f2b8a30))
+
+
+
+
+
+## [5.5.6](https://github.com/ykfe/ssr/compare/v5.5.5...v5.5.6) (2021-04-09)
+
+**Note:** Version bump only for package ssr-core-vue
+
+
+
+
+
+## [5.5.4](https://github.com/ykfe/ssr/compare/v5.5.2...v5.5.4) (2021-04-06)
+
+**Note:** Version bump only for package ssr-core-vue
+
+
+
+
+
 ## [5.5.2](https://github.com/ykfe/ssr/compare/v5.5.1...v5.5.2) (2021-04-05)
 
 **Note:** Version bump only for package ssr-core-vue