+++ packages/hoc-react/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.5.15](https://github.com/ykfe/ssr/compare/v5.5.14...v5.5.15) (2021-04-19)
+
+**Note:** Version bump only for package ssr-hoc-react
+
+
+
+
+
+## [5.5.14](https://github.com/ykfe/ssr/compare/v5.5.13...v5.5.14) (2021-04-17)
+
+
+### Features
+
+* update parseFeRoutes ([#51](https://github.com/ykfe/ssr/issues/51)) ([915e4cc](https://github.com/ykfe/ssr/commit/915e4cca22d0450b551fa6874c2b466648638374))
+
+
+
+
+
+## [5.5.13](https://github.com/ykfe/ssr/compare/v5.5.12...v5.5.13) (2021-04-15)
+
+**Note:** Version bump only for package ssr-hoc-react
+
+
+
+
+
+## [5.5.6](https://github.com/ykfe/ssr/compare/v5.5.5...v5.5.6) (2021-04-09)
+
+**Note:** Version bump only for package ssr-hoc-react
+
+
+
+
+
 # [5.5.0](https://github.com/ykfe/ssr/compare/v5.4.26...v5.5.0) (2021-04-05)
 
 **Note:** Version bump only for package ssr-hoc-react